Dickbutt Max HP is now Float instead of Int as all the other HP values are also floats

+++ Assets/Scripts/Objects/Player.cs
@@ -23,7 +23,7 @@ public class Player : MonoBehaviour
     private bool playerMovement = true;
     private int collectablesCount;
     private GameObject gameController;
-    private int maxHP = 100;
+    private float maxHP = 100.0f;
 
     void Start()
     {
